% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_coefs.R
\name{cleanup_coefs}
\alias{cleanup_coefs}
\title{Clean up names}
\usage{
cleanup_coefs(re_names, ran_coefs, se)
}
\arguments{
\item{re_names}{Names of the randome effects}

\item{ran_coefs}{The coefficients}

\item{se}{The standard errors of the coefficients}
}
\value{
A data frame with clean names
}
\description{
Only used internally to clean up names.
}
