% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fixed.R
\name{extract_fixed}
\alias{extract_fixed}
\title{Extract fixed effects}
\usage{
extract_fixed(model, ci_level = 0.95, ci_args = NULL, digits = 3,
  ...)
}
\arguments{
\item{model}{an lme4, glmmTMB, nlme, or brms model}

\item{ci_level}{confidence level < 1, typically above 0.90. A value of 0 will
not report it. Default is .95.}

\item{ci_args}{A list of additional arguments to the corresponding confint method. Default (\code{list(method = 'Wald')}) is to change the default CI method for speedier results.}

\item{digits}{Rounding. Default is 3.}

\item{...}{Other arguments to pass. Nothing at present.}
}
\value{
A data.frame with the fixed effects and associated statistics
}
\description{
Extract fixed effects
}
\details{
Essentially duplicates the \code{broom::tidy} approach with minor
  name changes.
}
\examples{
library(lme4)
library(mixedup)

lmer_mod <- lmer(Reaction ~ Days + (1 + Days | Subject), data = sleepstudy)

extract_fixed(lmer_mod)

}
\seealso{
\code{\link[broom:tidy.merMod]{tidy.merMod}}
}
