% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_random_effects.R
\name{extract_random_effects}
\alias{extract_random_effects}
\alias{extract_random_effects.merMod}
\alias{extract_random_effects.glmmTMB}
\alias{extract_random_effects.lme}
\alias{extract_random_effects.brmsfit}
\alias{extract_random_effects.stanreg}
\alias{extract_random_effects.gam}
\title{Extract random effects}
\usage{
extract_random_effects(model, re = NULL, ci_level = 0.95, digits = 3,
  component = NULL, ...)

\method{extract_random_effects}{merMod}(model, re = NULL,
  ci_level = 0.95, digits = 3, ...)

\method{extract_random_effects}{glmmTMB}(model, re = NULL,
  ci_level = 0.95, digits = 3, component = "cond", ...)

\method{extract_random_effects}{lme}(model, re = NULL, ci_level = NULL,
  digits = 3, ...)

\method{extract_random_effects}{brmsfit}(model, re = NULL,
  ci_level = 0.95, digits = 3, ...)

\method{extract_random_effects}{stanreg}(...)

\method{extract_random_effects}{gam}(model, re = NULL, ci_level = 0.95,
  digits = 3, ...)
}
\arguments{
\item{model}{An appropriate model. See details.}

\item{re}{The name of the grouping variable for the random effects. Default
is \code{NULL} to return all.}

\item{ci_level}{Where possible, confidence level < 1, typically above 0.90. A value of 0 will
not report it. Default is .95. Not applicable to nlme objects.}

\item{digits}{Rounding. Default is 3.}

\item{component}{Only applies to glmmTMB objects. Which of the three
components 'cond', 'zi' or 'other' to select for a glmmTMB model. Default
is 'cond'. Minimal testing on other options.}

\item{...}{Other arguments specific to the method. Unused at present.}
}
\value{
data frame of the random effects
}
\description{
Extracts the random effects and their standard errors.
}
\details{
Relative to \code{ranef} for the various packages, this just adds
  the standard errors and cluster ids as columns, and intervals.

\describe{
 \item{merMod}
 \item{glmmTMB}
 \item{brms}
 \item{nlme}
 \item{brms}
 \item{rstanarm}
 \item{mgcv}
}
}
\note{
\code{nlme} only provides the estimated random effect parameters, not their uncertainty, so it isn't provided.

\code{merMod} and \code{glmmTMB} results are based on the estimated conditional variances, i.e. \code{condvar = TRUE}.  This is likely an underestimate relative to brms results.

For \code{mgcv}, the `Vp` (Bayesian) estimated variance covariance matrix is used.
}
\examples{
library(lme4)
library(mixedup)

lmer_model <- lmer(Reaction ~ Days + (1 + Days | Subject), data = sleepstudy)
extract_random_effects(lmer_model)

}
\seealso{
\code{\link[lme4:ranef.merMod]{ranef.merMod}},
\code{\link[glmmTMB:ranef.glmmTMB]{ranef.glmmTMB}},
\code{\link[nlme:ranef.lme]{ranef.lme}},
\code{\link[brms:ranef.brmsfit]{ranef.brmsfit}},
\code{\link[mgcv:gamObject]{gamObject}},
\code{\link[mgcv:smooth.construct.re.smooth.spec]{smooth.construct.re.smooth.spec}}

\code{\link[lme4:ranef.merMod]{ranef.merMod}}, \code{\link[glmmTMB:ranef.glmmTMB]{ranef.glmmTMB}},
\code{\link[nlme:ranef.lme]{ranef.lme}}, \code{\link[brms:ranef.brmsfit]{ranef.brmsfit}}
}
