% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_quants.R
\name{find_quants}
\alias{find_quants}
\title{Find typical values of a random effect}
\usage{
find_quants(value, probs)
}
\arguments{
\item{value}{The random effect values}

\item{probs}{Desired probabilities passed to \code{quantile}}
}
\description{
Find typical values of a random effect
}
\details{
Finds indices of random effect values closest to a specific quantile
}
