% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_package_installed.R
\name{is_package_installed}
\alias{is_package_installed}
\title{Test package installation}
\usage{
is_package_installed(package)
}
\arguments{
\item{package}{Character string of the package name to test}
}
\description{
Test package installation
}
