% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedup-package.R
\docType{package}
\name{mixedup-package}
\alias{mixedup}
\alias{mixedup-package}
\title{mixedup: Miscellaneous functions for mixed models}
\description{
This provides tools for extracting various parts of mixed model output from commonly used packages in a consistent fashion. Results are produced in a usable form for visualization and table presentation. It currently provides a wide variety of support for lme4, nlme, glmmTMB, mgcv, brms, and rstanarm.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/m-clark/mixedup}
  \item Report bugs at \url{https://github.com/m-clark/mixedup/issues}
}

}
\author{
\strong{Maintainer}: Michael Clark \email{micl@umich.edu}

}
\keyword{internal}
