% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_model.R
\name{summarize_model}
\alias{summarize_model}
\alias{summarise_model}
\title{Summarize a mixed model}
\usage{
summarize_model(
  model,
  ci = TRUE,
  cor_re = FALSE,
  cor_fe = FALSE,
  digits = 3,
  component = "cond",
  ...
)

summarise_model(
  model,
  ci = TRUE,
  cor_re = FALSE,
  cor_fe = FALSE,
  digits = 3,
  component = "cond",
  ...
)
}
\arguments{
\item{model}{A supported model.}

\item{ci}{Whether to include a 95\% uncertainty interval. Default is TRUE.}

\item{cor_re}{Whether to include the correlations of the random effects.
Default is FALSE.}

\item{cor_fe}{Whether to include the correlations of the fixed effects.
Default is FALSE.}

\item{digits}{Digits to display.}

\item{component}{For glmmTMB objects, which of the three components 'cond',
'zi' or 'other' to select. Default is cond. Minimal testing on other
options.}

\item{...}{Not used at present. May allow models specific functionality.}
}
\description{
Summarize a mixed model
}
\examples{
library(lme4)
library(mixedup)

lmer_mod <- lmer(Reaction ~ Days + (1 + Days | Subject), data = sleepstudy)
summarize_model(lmer_mod)


}
