% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cor_structure.R
\name{extract_cor_structure}
\alias{extract_cor_structure}
\alias{extract_cor_structure.lme}
\title{Extract correlation structure}
\usage{
extract_cor_structure(model, ...)

\method{extract_cor_structure}{lme}(model, ...)
}
\arguments{
\item{model}{An appropriate mixed model}

\item{...}{Other arguments appropriate to the method}
}
\value{
A vector of the estimates.
}
\description{
Extract residual correlation structure for nlme, and potentially other models.
}
\details{
For models with correlation, i.e. that contain something like
corAR1(form = ~time).  For more see this
\href{https://bbolker.github.io/mixedmodels-misc/notes/corr_braindump.html}{'braindump'
from Ben Bolker}.
}
