% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fixed_effects.R
\name{extract_fixed_effects}
\alias{extract_fixed_effects}
\alias{extract_fixed_effects.merMod}
\alias{extract_fixed_effects.glmmTMB}
\alias{extract_fixed_effects.lme}
\alias{extract_fixed_effects.brmsfit}
\alias{extract_fixed_effects.gam}
\title{Extract fixed effects}
\usage{
extract_fixed_effects(
  model,
  ci_level = 0.95,
  ci_args = NULL,
  digits = 3,
  exponentiate = FALSE,
  ...
)

\method{extract_fixed_effects}{merMod}(
  model,
  ci_level = 0.95,
  ci_args = list(method = "Wald"),
  digits = 3,
  exponentiate = FALSE,
  ...
)

\method{extract_fixed_effects}{glmmTMB}(
  model,
  ci_level = 0.95,
  ci_args = NULL,
  digits = 3,
  exponentiate = FALSE,
  component = "cond",
  ...
)

\method{extract_fixed_effects}{lme}(
  model,
  ci_level = 0.95,
  ci_args = list(method = "Wald"),
  digits = 3,
  exponentiate = FALSE,
  ...
)

\method{extract_fixed_effects}{brmsfit}(
  model,
  ci_level = 0.95,
  ci_args = NULL,
  digits = 3,
  exponentiate = FALSE,
  component = NULL,
  ...
)

\method{extract_fixed_effects}{gam}(
  model,
  ci_level = 0.95,
  ci_args = list(method = "Wald"),
  digits = 3,
  exponentiate = FALSE,
  ...
)
}
\arguments{
\item{model}{an lme4, glmmTMB, nlme, mgcv, or brms model}

\item{ci_level}{confidence level < 1, typically above 0.90. A value of 0 will
not report it (except for gam objects, which will revert to .95 due to
\code{gam.vcomp}). Default is .95.}

\item{ci_args}{Additional arguments to the corresponding confint method.}

\item{digits}{Rounding. Default is 3.}

\item{exponentiate}{Exponentiate the fixed-effect coefficient estimates and
confidence intervals (common for logistic regression). If TRUE, also scales
the standard errors by the exponentiated coefficient, transforming them to
the new scale.}

\item{...}{Other stuff to pass to the corresponding function. Unused/tested}

\item{component}{For glmmTMB objects, which of the three components 'cond' or
'zi' to select. Default is 'cond'.  For brmsfit objects, this can filter
results to a certain part of the output, e.g. 'sigma' or 'zi' of
distributional models, or a specific outcome of a multivariate model.  In
this case \code{component} is a regular expression that begins parameters
of the output.}
}
\value{
A data.frame with the fixed effects and associated statistics.
}
\description{
Extract fixed effects
}
\details{
Essentially duplicates the \code{broom::tidy} approach with minor
name changes.  The package may or may not provide p-values by default.
}
\note{
For nlme, this is just a multiplier based on the estimated standard
error and critical value for the \code{ci_level}.
}
\examples{
library(lme4)
library(mixedup)

lmer_mod <- lmer(Reaction ~ Days + (1 + Days | Subject), data = sleepstudy)

extract_fixed_effects(lmer_mod)
}
\seealso{
\code{\link[broom:tidy.merMod]{tidy.merMod}},
\code{\link[broom.mixed:tidy.glmmTMB]{tidy.glmmTMB}},
\code{\link[broom.mixed:tidy.brmsfit]{tidy.brmsfit}}
}
