% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_het_var.R
\name{extract_het_var}
\alias{extract_het_var}
\alias{extract_het_var.lme}
\alias{extract_het_var.glmmTMB}
\title{Extract heterogeneous variances}
\usage{
extract_het_var(model, digits = 3, ...)

\method{extract_het_var}{lme}(model, digits = 3, ...)

\method{extract_het_var}{glmmTMB}(model, digits = 3, ...)
}
\arguments{
\item{model}{An appropriate mixed model.}

\item{digits}{Rounding. Default is 3.}

\item{...}{Other arguments appropriate to the method.}
}
\value{
A vector of the estimates on the variance scale.
}
\description{
Extract heterogeneous variances for nlme, and potentially other models.
}
\details{
For models with heterogeneous variance, i.e. that contain something
like \code{varIdent(form = ~1|Group)}, this returns a more presentable version
of the estimates. Only tested with the \code{varIdent} case.

For glmmTMB, this serves as a wrapper for \link{extract_cor_structure} with for the
'diag' option.  See that function for details.
}
\examples{
library(nlme)
library(mixedup)

model <- lme(distance ~ age + Sex, data = Orthodont, random = ~ 1|Subject,
weights = varIdent(form = ~ 1 | Sex))

summary(model)

extract_het_var(model)

library(glmmTMB)

# does not get the same estimates as nlme, but would get similar if modeled
# using dispersion approach.
model <-
  glmmTMB(distance ~ age + Sex + (1 | Subject) + diag(Sex + 0 | Subject),
          data = Orthodont)

extract_het_var(model)

}
\seealso{
Other extract: 
\code{\link{extract_cor_structure}()},
\code{\link{extract_fixed_effects}()},
\code{\link{extract_model_data}()},
\code{\link{extract_random_coefs}()},
\code{\link{extract_random_effects}()},
\code{\link{extract_vc}()}
}
\concept{extract}
