% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_het_var.R
\name{extract_het_var}
\alias{extract_het_var}
\title{extract_nlme_variances}
\usage{
extract_het_var(model, ...)
}
\arguments{
\item{model}{An appropriate mixed model}

\item{...}{Other arguments appropriate to the method}
}
\value{
A vector of the estimates on the variance scale.
}
\description{
Because nlme won't do it for you.
}
\details{
For models with heterogeneous variance, i.e. that contain something
  like varIdent(form = ~1|Group), nlme returns a result in the summary
  regarding the variances that many do not know what to do with, nor likely
  would be what they would want to report. I rarely do these models with
  nlme, and have only played around with the varIdent case.   This function '
  will save me a little trouble in demos I do, but I have no plans to do any
  more with it.
}
\examples{
library(nlme)
library(mixedup)

model <- lme(distance ~ age + Sex, data = Orthodont, random = ~ 1|Subject,
weights=varIdent(form = ~1|Sex))

summary(model)

extract_het_var(model)

}
