% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_model_data.R
\name{extract_model_data}
\alias{extract_model_data}
\title{Extract model data}
\usage{
extract_model_data(model)
}
\arguments{
\item{model}{The mixed model.}
}
\value{
A tibble/dataframe of the data used in the model.
}
\description{
This is mostly just a wrapper for \code{model.frame}.
}
\note{
For whatever reason, \code{nlme} class objects do not save the model data,
so this will throw an error stating as much. \code{lme} objects do save the data
so are fine.
}
\examples{
library(lme4)
library(mixedup)

lmer_1 <- lmer(Reaction ~ Days + (1 | Subject), data = sleepstudy)
extract_model_data(lmer_1)


}
\seealso{
\link{model.frame}

Other extract: 
\code{\link{extract_cor_structure}()},
\code{\link{extract_fixed_effects}()},
\code{\link{extract_het_var}()},
\code{\link{extract_random_coefs}()},
\code{\link{extract_random_effects}()},
\code{\link{extract_vc}()}
}
\concept{extract}
