% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_random_coef.R
\name{extract_random_coef}
\alias{extract_random_coef}
\title{Extract random coefficients and their variances}
\usage{
extract_random_coef(model, which_re = NULL, zi = FALSE)
}
\arguments{
\item{model}{A merMod or glmmTMB object}

\item{which_re}{The name of the grouping variable desired.}

\item{zi}{For a zero-inflated glmmTMB model, which part of the model you want random coefficients for?}
}
\value{
A data frame
}
\description{
Extract random coefficients and their variances
}
\examples{
library(lme4)

}
