% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_random_coef.R
\name{extract_random_coef}
\alias{extract_random_coef}
\alias{extract_random_coef.merMod}
\alias{extract_random_coef.glmmTMB}
\alias{extract_random_coef.lme}
\alias{extract_random_coef.brmsfit}
\title{Extract random coefficients and their variances}
\usage{
extract_random_coef(model, re = NULL, ci_level = 0.95, digits = 3,
  ...)

\method{extract_random_coef}{merMod}(model, re = NULL, ci_level = 0.95,
  digits = 3, ...)

\method{extract_random_coef}{glmmTMB}(model, re = NULL,
  ci_level = 0.95, digits = 3, ...)

\method{extract_random_coef}{lme}(model, re = NULL, digits = 3, ...)

\method{extract_random_coef}{brmsfit}(model, re = NULL,
  ci_level = 0.95, digits = 3, ...)
}
\arguments{
\item{model}{A merMod or glmmTMB object}

\item{re}{The name of the grouping variable for the random effects.}

\item{ci_level}{Where possible, confidence level < 1, typically above 0.90. A value of 0 will
not report it. Default is .95.}

\item{digits}{Rounding. Default is 3.}

\item{...}{Other arguments specific to the method. Unused at present.}

\item{component}{Which of the three components 'cond', 'zi' or 'other' to
select for a glmmTMB model. Default is 'cond'. Minimal testing on other
options.}
}
\value{
A data frame of the random coefficients and their standard errors.
}
\description{
Extract random coefficients and their variances
}
\details{
Returns a data frame with random coefficients, a.k.a. random
  intercepts and random slopes, and their standard errors. Note that the
  standard errors assume independence of the conditional variance and the
  fixed-effects variance, thus the standard errors are the sum of variances
  for the fixed and random effects. See Bolker's demo
  \href{https://stackoverflow.com/questions/26198958/extracting-coefficients-and-their-standard-error-from-lme}{here}
  and additional discussion at the
  \href{https://bbolker.github.io/mixedmodels-misc/glmmFAQ.html#confidence-intervals-on-conditional-meansblupsrandom-effects}{GLMM
  FAQ}. This assumption may not be appropriate.


\code{nlme} only provides the coefficients no estimated variance, so this
doesn't add to what you get from basic functionality for those models.  In
addition, nlme adds all random effects to the fixed effects, whereas
\code{lme4} and others only add the effects requested.
}
\examples{
library(lme4)
lmer_1 <- lmer(Reaction ~ Days + (1 | Subject), data = sleepstudy)
extract_random_coef(lmer_1, re = 'Subject')
library(glmmTMB)
tmb_1 <- glmmTMB(Reaction ~ Days + (1 | Subject), data = sleepstudy)
extract_random_coef(tmb_1, re = 'Subject')

}
