% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_random_coef.R
\name{extract_random_coef}
\alias{extract_random_coef}
\title{Extract random coefficients and their variances}
\usage{
extract_random_coef(model, re = NULL, zi = FALSE)
}
\arguments{
\item{model}{A merMod or glmmTMB object}

\item{re}{The name of the grouping variable for the random effects.}

\item{zi}{For a zero-inflated glmmTMB model, which part of the model you want
random coefficients for? Default is false, as typical mixed models are not
zero-inflated models. This will use the 'conditional model' results.}
}
\value{
A data frame of the random coefficients and their standard errors.
}
\description{
Extract random coefficients and their variances
}
\details{
Returns a data frame with random coefficients, a.k.a. random
intercepts and random slopes, and their standard errors.   The standard
errors are the sum of variances for the fixed and random effects. See
Bolker's demo
\href{https://stackoverflow.com/questions/26198958/extracting-coefficients-and-their-standard-error-from-lme}{here}.
}
\examples{
library(lme4)
lmer_1 <- lmer(Reaction ~ Days + (1 | Subject), data = sleepstudy)
extract_random_coef(lmer_1, re = 'Subject')
library(glmmTMB)
tmb_1 <- glmmTMB(Reaction ~ Days + (1 | Subject), data = sleepstudy)
extract_random_coef(tmb_1, re = 'Subject')

}
