% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_random_coefs.R
\name{extract_random_coefs}
\alias{extract_random_coefs}
\alias{extract_random_coefs.merMod}
\alias{extract_random_coefs.glmmTMB}
\alias{extract_random_coefs.lme}
\alias{extract_random_coefs.brmsfit}
\alias{extract_random_coefs.stanreg}
\alias{extract_random_coefs.gam}
\title{Extract random coefficients and their variances}
\usage{
extract_random_coefs(
  model,
  re = NULL,
  ci_level = 0.95,
  digits = 3,
  component = NULL,
  ...
)

\method{extract_random_coefs}{merMod}(model, re = NULL, ci_level = 0.95, digits = 3, ...)

\method{extract_random_coefs}{glmmTMB}(
  model,
  re = NULL,
  ci_level = 0.95,
  digits = 3,
  component = "cond",
  ...
)

\method{extract_random_coefs}{lme}(model, re = NULL, ci_level = NULL, digits = 3, ...)

\method{extract_random_coefs}{brmsfit}(
  model,
  re = NULL,
  ci_level = 0.95,
  digits = 3,
  component = NULL,
  ...
)

\method{extract_random_coefs}{stanreg}(
  model,
  re = NULL,
  ci_level = 0.95,
  digits = 3,
  component = NULL,
  ...
)

\method{extract_random_coefs}{gam}(model, re = NULL, ci_level = 0.95, digits = 3, ...)
}
\arguments{
\item{model}{A merMod, nlme, brms, or glmmTMB object}

\item{re}{The name of the grouping variable for the random effects.}

\item{ci_level}{Where possible, confidence level < 1, typically above 0.90. A value of 0 will
not report it. Default is .95. Not applicable to nlme objects.}

\item{digits}{Rounding. Default is 3.}

\item{component}{For glmmTMB objects, which of the two components 'cond' or
'zi' to select. Default is 'cond'. For brmsfit objects, this can filter
results to a certain part of the output, e.g. 'sigma' or 'zi' of
distributional models, or a specific outcome of a multivariate model.  In
this case \code{component} is a regular expression that ends the name of the
parameters of the output (e.g. '__component').}

\item{...}{Other arguments specific to the method. Unused at present.}
}
\value{
A data frame of the random coefficients and their standard errors.
}
\description{
Fixed effect + random effects.
}
\details{
Returns a data frame with random coefficients, a.k.a. random
intercepts and random slopes, and their standard errors. Note that the
standard errors assume independence of the conditional variance and the
fixed-effects variance, thus the standard errors are the sum of variances
for the respective fixed and random effects. See Bolker's demo
\href{https://stackoverflow.com/questions/26198958/extracting-coefficients-and-their-standard-error-from-lme}{here}
and additional discussion at the
\href{https://bbolker.github.io/mixedmodels-misc/glmmFAQ.html#confidence-intervals-on-conditional-meansblupsrandom-effects}{GLMM
FAQ}. As noted there, this assumption may not be appropriate, and if you
are really interested in an accurate uncertainty estimate you should
probably use \code{brms}.

For more complex models that include multiple outcomes/categories, this
function likely will not work at present, as naming conventions are not
consistent. I will possibly be able to update this in the future.

The \code{nlme} package only provides the coefficients no estimated variance, so this
function doesn't add to what you get from basic functionality for those
models.  In addition, \code{nlme} adds all random effects to the fixed effects,
whereas \code{lme4} and others only add the effects requested.
}
\examples{
library(lme4)
library(mixedup)

lmer_1 <- lmer(Reaction ~ Days + (1 | Subject), data = sleepstudy)
extract_random_coefs(lmer_1, re = 'Subject')
library(glmmTMB)
tmb_1 <- glmmTMB(Reaction ~ Days + (1 | Subject), data = sleepstudy)
extract_random_coefs(tmb_1, re = 'Subject')

}
\seealso{
Other extract: 
\code{\link{extract_cor_structure}()},
\code{\link{extract_fixed_effects}()},
\code{\link{extract_het_var}()},
\code{\link{extract_model_data}()},
\code{\link{extract_random_effects}()},
\code{\link{extract_vc}()}
}
\concept{extract}
