% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_random_coefs.R
\name{extract_random_coefs}
\alias{extract_random_coefs}
\alias{extract_random_coefs.merMod}
\alias{extract_random_coefs.glmmTMB}
\alias{extract_random_coefs.lme}
\alias{extract_random_coefs.brmsfit}
\alias{extract_random_coefs.gam}
\title{Extract random coefficients and their variances}
\usage{
extract_random_coefs(
  model,
  re = NULL,
  ci_level = 0.95,
  digits = 3,
  component = NULL,
  ...
)

\method{extract_random_coefs}{merMod}(model, re = NULL, ci_level = 0.95, digits = 3, ...)

\method{extract_random_coefs}{glmmTMB}(
  model,
  re = NULL,
  ci_level = 0.95,
  digits = 3,
  component = "cond",
  ...
)

\method{extract_random_coefs}{lme}(model, re = NULL, ci_level = NULL, digits = 3, ...)

\method{extract_random_coefs}{brmsfit}(
  model,
  re = NULL,
  ci_level = 0.95,
  digits = 3,
  component = NULL,
  ...
)

\method{extract_random_coefs}{gam}(model, re = NULL, ci_level = 0.95, digits = 3, ...)
}
\arguments{
\item{model}{A merMod, nlme, brms, or glmmTMB object}

\item{re}{The name of the grouping variable for the random effects.}

\item{ci_level}{Where possible, confidence level < 1, typically above 0.90. A value of 0 will
not report it. Default is .95. Not applicable to nlme objects.}

\item{digits}{Rounding. Default is 3.}

\item{component}{Only applies to \code{glmmTMB} objects. Which of the three
components 'cond', 'zi' or 'other' to select for a glmmTMB model. Default
is 'cond'. Minimal testing on other options.}

\item{...}{Other arguments specific to the method. Unused at present.}
}
\value{
A data frame of the random coefficients and their standard errors.
}
\description{
Extract random coefficients and their variances
}
\details{
Returns a data frame with random coefficients, a.k.a. random
intercepts and random slopes, and their standard errors. Note that the
standard errors assume independence of the conditional variance and the
fixed-effects variance, thus the standard errors are the sum of variances
for the respective fixed and random effects. See Bolker's demo
\href{https://stackoverflow.com/questions/26198958/extracting-coefficients-and-their-standard-error-from-lme}{here}
and additional discussion at the
\href{https://bbolker.github.io/mixedmodels-misc/glmmFAQ.html#confidence-intervals-on-conditional-meansblupsrandom-effects}{GLMM
FAQ}. As noted there, this assumption may not be appropriate, and if you
are really interested in an accurate uncertainty estimate you should
probably use brms.

\code{nlme} only provides the coefficients no estimated variance, so this
function doesn't add to what you get from basic functionality for those
models.  In addition, nlme adds all random effects to the fixed effects,
whereas \code{lme4} and others only add the effects requested.
}
\examples{
library(lme4)
library(mixedup)

lmer_1 <- lmer(Reaction ~ Days + (1 | Subject), data = sleepstudy)
extract_random_coefs(lmer_1, re = 'Subject')
library(glmmTMB)
tmb_1 <- glmmTMB(Reaction ~ Days + (1 | Subject), data = sleepstudy)
extract_random_coefs(tmb_1, re = 'Subject')

}
\seealso{
\code{\link[mixedup]{extract_random_effects}},
\code{\link[mixedup]{extract_fixed_effects}}
}
