% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_random_effects.R
\name{extract_random_effects}
\alias{extract_random_effects}
\title{Extract random effects}
\usage{
extract_random_effects(model, re = NULL, component = "cond",
  digits = 3)
}
\arguments{
\item{model}{A merMod or glmmTMB object}

\item{re}{The name of the grouping variable for the random effects. Default
is \code{NULL} to return all.}

\item{component}{Which of the three components 'cond', 'zi' or 'other' to
select for a glmmTMB model. Default is 'cond'. Minimal testing on other
options.}

\item{digits}{Rounding. Default is 3.}
}
\value{
data frame of the random effects
}
\description{
Extracts the random effects and their standard errors.
}
\details{
Relative to \code{ranef} for the various packages, this just adds
  the standard errors and cluster ids as columns, and intervals for brmsfit objects.

\code{nlme} only provides the estimated random effect parameters, not their uncertainty, so it isn't provided.
}
\examples{
library(lme4)
lmer_2 <- lmer(Reaction ~ Days + (1 + Days | Subject), data = sleepstudy)
extract_random_effects(lmer_2)

}
\seealso{
\code{\link[lme4:ranef]{ranef.merMod}}, \code{\link[glmmTMB:ranef]{ranef.glmmTMB}},
\code{\link[nlme:ranef]{ranef.lme}}, \code{\link[brms:ranef]{ranef.brmsfit}}
}
