% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_vc.R
\name{extract_vc}
\alias{extract_vc}
\alias{extract_vc.merMod}
\alias{extract_vc.glmmTMB}
\title{Extract variance components}
\usage{
extract_vc(model, ci_level = 0.95, ci_args = NULL, ci_scale = "sd",
  component = "cond", show_cor = FALSE, digits = 3, ...)

\method{extract_vc}{merMod}(model, ci_level = 0.95, ci_args = NULL,
  ci_scale = "sd", component, show_cor = FALSE, digits = 3, ...)

\method{extract_vc}{glmmTMB}(model, ci_level = 0.95, ci_args = NULL,
  ci_scale = "sd", component = "cond", show_cor = FALSE,
  digits = 3, ...)
}
\arguments{
\item{model}{an lme4 or glmmTMB model}

\item{ci_level}{confidence level < 1, typically above 0.90. A value of 0 will
not report it. Default is .95.}

\item{ci_args}{Additional arguments to the corresponding confint method.}

\item{ci_scale}{A character string of 'sd' or 'var' to note the scale of the
interval estimate.  Default is 'sd'.
at present.}

\item{component}{For glmmTMB objects, which of the three components 'cond',
'zi' or 'other' to select. Default is cond. Minimal testing on other
options.}

\item{show_cor}{Return the intercept/slope correlations. Default is
\code{FALSE}.}

\item{digits}{Rounding. Default is 3.}

\item{...}{Other stuff to pass to the corresponding function. Unused/tested}
}
\value{
A data frame with output for variance components or list that also
contains the correlations of the random effects.
}
\description{
Right now this only works for lme4 and glmmTMB model objects (i.e.
  \code{merMod}).
}
\details{
Returns a more usable (my opinion) version of variance components
  estimates including variance, standard deviation, the confidence interval
  for either, the relative proportion of variance, and all in a data frame
  with names that are clean and easy to use.
}
\examples{
library(lme4)

lmer_mod <- lmer(Reaction ~ Days + (1 + Days | Subject), data = sleepstudy)

extract_vc(lmer_mod)
extract_vc(lmer_mod, ci_scale = 'var')

}
\seealso{
\code{\link{confint.merMod}}, \code{\link{VarCorr.merMod}},
  \code{\link{confint.glmmTMB}}, \code{\link{VarCorr.glmmTMB}},
  \code{\link{intervals}}, \code{\link{VarCorr.lme}}
}
