% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_vc.R
\name{extract_vc}
\alias{extract_vc}
\alias{extract_vc.merMod}
\alias{extract_vc.glmmTMB}
\alias{extract_vc.lme}
\alias{extract_vc.brmsfit}
\alias{extract_vc.stanreg}
\alias{extract_vc.gam}
\alias{extract_variance_components}
\alias{extract_VarCorr}
\title{Extract variance components}
\usage{
extract_vc(
  model,
  ci_level = 0.95,
  ci_args = NULL,
  ci_scale = "sd",
  show_cor = FALSE,
  digits = 3,
  component = "cond",
  ...
)

\method{extract_vc}{merMod}(
  model,
  ci_level = 0.95,
  ci_args = NULL,
  ci_scale = "sd",
  show_cor = FALSE,
  digits = 3,
  ...
)

\method{extract_vc}{glmmTMB}(
  model,
  ci_level = 0.95,
  ci_args = NULL,
  ci_scale = "sd",
  show_cor = FALSE,
  digits = 3,
  component = "cond",
  ...
)

\method{extract_vc}{lme}(
  model,
  ci_level = 0.95,
  ci_args = NULL,
  ci_scale = "sd",
  show_cor = FALSE,
  digits = 3,
  ...
)

\method{extract_vc}{brmsfit}(
  model,
  ci_level = 0.95,
  ci_args = NULL,
  ci_scale = "sd",
  show_cor = FALSE,
  digits = 3,
  component = NULL,
  ...
)

\method{extract_vc}{stanreg}(
  model,
  ci_level = 0.95,
  ci_args = NULL,
  ci_scale = "sd",
  show_cor = FALSE,
  digits = 3,
  component = NULL,
  ...
)

\method{extract_vc}{gam}(
  model,
  ci_level = 0.95,
  ci_args = NULL,
  ci_scale = "sd",
  show_cor = FALSE,
  digits = 3,
  ...
)

extract_variance_components(
  model,
  ci_level = 0.95,
  ci_args = NULL,
  ci_scale = "sd",
  show_cor = FALSE,
  digits = 3,
  component = "cond",
  ...
)

extract_VarCorr(
  model,
  ci_level = 0.95,
  ci_args = NULL,
  ci_scale = "sd",
  show_cor = FALSE,
  digits = 3,
  component = "cond",
  ...
)
}
\arguments{
\item{model}{An lme4, glmmTMB, nlme, mgcv, or brms model.}

\item{ci_level}{Confidence level < 1, typically above 0.90. A value of 0 will
not report it (except for gam objects, which will revert to .95 due to
\code{gam.vcomp}). Default is .95.}

\item{ci_args}{Additional arguments to the corresponding confint method.}

\item{ci_scale}{A character string of 'sd' or 'var' to note the scale of the
interval estimate.  Default is 'sd'. at present.}

\item{show_cor}{Return the intercept/slope correlations as a separate list
element. Default is \code{FALSE}.}

\item{digits}{Rounding. Default is 3.}

\item{component}{For glmmTMB objects, which of the three components 'cond' or
'zi' to select. Default is 'cond'.  For brmsfit (and experimentally,
rstanarm) objects, this can filter results to a certain part of the output,
e.g. 'sigma' or 'zi' of distributional models, or a specific outcome of a
multivariate model.  In this case \code{component} is a regular expression
that begins parameters
of the output.}

\item{...}{Other stuff to pass to the corresponding function. Unused/tested}
}
\value{
A data frame with output for variance components, or list that also
contains the correlations of the random effects.
}
\description{
This has functionality for simpler models from \code{lme4},
\code{glmmTMB}, \code{nlme}, and \code{brms}.
}
\details{
Returns a more usable (my opinion) version of variance components
estimates including variance, standard deviation, the confidence interval
for either, the relative proportion of variance, and all in a data frame
with names that are clean and easy to use.

\code{extract_variance_components} and \code{extract_VarCorr} are aliases.
}
\note{
Right now, there are several issues with getting confidence intervals
for \code{glmmTMB} objects
(\href{https://github.com/glmmTMB/glmmTMB/issues/571}{for example, which is
actually not resolved}).  If you get an error, you should check by running
\code{confint(my_tmb_model)} before posting an issue.  While I've attempted some
minor hacks to deal with some of them, it stands to reason that if the
\code{glmmTMB} functionality doesn't work, this function won't either.
}
\examples{
library(lme4)
library(mixedup)

lmer_mod <- lmer(Reaction ~ Days + (1 + Days | Subject), data = sleepstudy)

extract_vc(lmer_mod)
extract_vc(lmer_mod, ci_scale = 'var')

}
\seealso{
\code{\link[lme4:confint.merMod]{lme4::confint.merMod()}},
\code{\link[lme4:VarCorr]{lme4::VarCorr()}},
\code{\link[glmmTMB:VarCorr.glmmTMB]{glmmTMB::VarCorr()}},
\code{\link[nlme:intervals]{nlme::intervals()}},
\code{\link[nlme:VarCorr]{nlme::VarCorr()}},
\code{\link[brms:VarCorr.brmsfit]{brms::VarCorr()}},
\code{\link[rstanarm:stanreg-methods]{rstanarm::VarCorr()}},
\code{\link[mgcv:gam.vcomp]{mgcv::gam.vcomp()}}

Other extract: 
\code{\link{extract_cor_structure}()},
\code{\link{extract_fixed_effects}()},
\code{\link{extract_het_var}()},
\code{\link{extract_model_data}()},
\code{\link{extract_random_coefs}()},
\code{\link{extract_random_effects}()}
}
\concept{extract}
